Renamed package of Relay from 'util' to 'api'.
