+++ src/main/java/com/ircclouds/irc/api/negotiators/SaslNegotiator.java
@@ -6,7 +6,7 @@ import com.ircclouds.irc.api.commands.*;
 import com.ircclouds.irc.api.domain.messages.ServerNumericMessage;
 import com.ircclouds.irc.api.domain.messages.interfaces.*;
 import com.ircclouds.irc.api.listeners.VariousMessageListenerAdapter;
-import com.ircclouds.irc.api.negotiators.util.Relay;
+import com.ircclouds.irc.api.negotiators.api.Relay;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import org.slf4j.Logger;
